installer: consolidate binassets and assets, move tmpl rendering there
