+++ installer/assets/doc.go
@@ -1,3 +1,3 @@
 // Stub file to keep glide happy.
 
-package binassets
+package assets
