+++ installer/server/terraform.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"path/filepath"
 	"time"
-	
+
 	"golang.org/x/crypto/bcrypt"
 	"golang.org/x/net/context"
